depth grayscale

+++ run_video.py
@@ -80,7 +80,8 @@ if __name__ == '__main__':
             depth = F.interpolate(depth[None], (frame_height, frame_width), mode='bilinear', align_corners=False)[0, 0]
             depth = (depth - depth.min()) / (depth.max() - depth.min()) * 255.0
 
-            depth = depth.cpu().numpy().astype(np.uint8)
+            depth = depth.cpu().numpy()
+            # depth = depth.cpu().numpy().astype(np.uint8)
             # depth_color = cv2.applyColorMap(depth, cv2.COLORMAP_INFERNO)
 
             out.write(depth)
